+++ src/utils/excel.ts
@@ -72,6 +72,7 @@ export const excelExport = (data: any) => {
     addSheet('Afboekingen',           data.afboekingen)
     addSheet('Klanten',               data.klanten)
     addSheet('GistMetingen',          data.gist_metingen)
+    addSheet('CarbonatieSessies',     data.carbonatie_sessies)
     addSheet('KapitaalBoekingen',     data.kapitaal_boekingen)
     addSheet('EADDocumenten',         data.ead_documenten)
     addSheet('Inventarisaties',       data.inventarisaties)
@@ -186,6 +187,7 @@ export const excelImport = (file: File, cb: (data: any) => void, onError?: () =>
         afboekingen:                  parse('Afboekingen'),
         klanten:                      parse('Klanten'),
         gist_metingen:                parse('GistMetingen'),
+        carbonatie_sessies:           parse('CarbonatieSessies'),
         kapitaal_boekingen:           parse('KapitaalBoekingen'),
         ead_documenten:               parse('EADDocumenten'),
         inventarisaties:              parse('Inventarisaties'),